+++ ex_webrtc/integration_test/test_videoroom/lib/test_videoroom/room.ex
@@ -57,7 +57,6 @@ defmodule TestVideoroom.Room do
 
     endpoint = %ExWebRTC{
       rtc_engine: state.rtc_engine,
-      ice_port_range: state.ice_port_range,
       video_codec: :VP8,
       event_serialization: Application.fetch_env!(:test_videoroom, :event_serialization)
     }