+++ ex_webrtc/CHANGELOG.md
@@ -15,3 +15,4 @@
 * Add manual variant switching [#20](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/20)
 * Fix simulcast tests [#22](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/22)
 * Fix layer switching [#24](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/24)
+* Allow for configuring ice servers [#25](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/25)